// types/pdf.ts
export type StudentResultPDFtype = {
  name: string;
  fatherName: string;
  motherName: string;
  rollNo: string;
  regNo: string;
  result: string;
  branch: string;
  photoUrl: string; // Full URL of the student's photo from Cloudinary
  logoPath: string; // Path to the logo in the public folder
};